real/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
httpx==0.24.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0